# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "imperia-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Mapped server"
        },
        {
            "url": "http://localhost:8080",
            "description": "Current server"
        },
        {
            "url": "https://imperia-api.herokuapp.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register user.",
                "description": "Register user.",
                "operationId": "register",
                "requestBody": {
                    "description": "Register user request object.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User has been successfully registered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Index categories.",
                "operationId": "indexCategories",
                "parameters": [
                    {
                        "name": "page[size]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageSize"
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageNumber"
                        }
                    },
                    {
                        "name": "filter[target]",
                        "in": "query",
                        "description": "Target class morph slug. Examples: `products`, `tickets`, `services`, `spaces`",
                        "schema": {
                            "type": "string"
                        },
                        "example": "products"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Index categories response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IndexCategoryResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Show category by id.",
                "operationId": "showCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the category.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show category response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowCategoryResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RegisterResponse": {
                "description": "Register user response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/UserAndTokenResponse"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
...
            "UnauthenticatedResponse": {
                "description": "Unauthenticated response object.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
            "ValidationErrorsResponse": {
                "description": "Validation errors response object.",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "example": "The example field is required."
                        }
                    }
                },
                "type": "object"
            },
            "UserAndTokenResponse": {
                "description": "User and token response object.",
                "required": [
                    "token",
                    "user"
                ],
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|ARDGny3pfUiPpxaXCKtW1w5B7qqcqAYehtRTVA849X3p"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "IndexCategoryResponse": {
                "description": "Index categories response object.",
                "required": [
                    "data",
                    "meta",
                    "message"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "ShowCategoryResponse": {
                "description": "Show category response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
...
            "PageSize": {
                "description": "Max number of results on page.",
                "type": "integer",
                "example": 25
            },
            "PageNumber": {
                "description": "Page number to be returned.",
                "type": "integer",
                "example": 1
            },
...
            "RegisterRequest": {
                "description": "Register user request",
                "required": [
                    "name",
                    "surnane",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@email.com"
                    },
                    "phone": {
                        "nullable": "true",
                        "type": "string",
                        "example": "null"
                    },
                    "password": {
                        "type": "string",
                        "example": "pa$$w0rd"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "pa$$w0rd"
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "description": "Category resource object",
                "required": [
                    "id",
                    "type",
                    "slug",
                    "target",
                    "title",
                    "description",
                    "media",
                    "default_media"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "pizza"
                    },
                    "type": {
                        "type": "string",
                        "example": "categories"
                    },
                    "target": {
                        "nullable": "true",
                        "type": "string",
                        "example": "products"
                    },
                    "title": {
                        "type": "string",
                        "example": "Pizza"
                    },
                    "description": {
                        "nullable": "true",
                        "example": "Some text..."
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    },
                    "default_media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    }
                },
                "type": "object"
            },
...
            "Media": {
                "description": "Media resource object",
                "required": [
                    "id",
                    "type",
                    "name",
                    "title",
                    "url",
                    "path",
                    "folder",
                    "private"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "media"
                    },
                    "name": {
                        "type": "string",
                        "example": "mojito-1645908379CXJyz.jpg"
                    },
                    "title": {
                        "type": "string",
                        "example": "Mojito"
                    },
                    "url": {
                        "type": "string",
                        "example": "http://localhost/storage/mojito-1645908379CXJyz.jpg"
                    },
                    "path": {
                        "type": "string",
                        "example": "/mojito-1645908379CXJyz.jpg"
                    },
                    "folder": {
                        "type": "string",
                        "example": "/"
                    },
                    "private": {
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "type": "object"
            },
...
            "Links": {
                "description": "Pagination meta links object.",
                "required": [
                    "first",
                    "prev",
                    "self",
                    "next",
                    "last"
                ],
                "properties": {
                    "first": {
                        "type": "string"
                    },
                    "prev": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    },
                    "self": {
                        "type": "string"
                    },
                    "next": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    },
                    "last": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "description": "Pagination meta object.",
                "required": [
                    "from",
                    "to",
                    "total",
                    "path",
                    "per_page",
                    "current_page",
                    "last_page",
                    "links"
                ],
                "properties": {
                    "from": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 30
                    },
                    "path": {
                        "nullable": "true",
                        "type": "integer",
                        "example": null
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "links": {
                        "$ref": "#/components/schemas/Links"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "description": "User resource object",
                "required": [
                    "id",
                    "type",
                    "customer_id",
                    "name",
                    "email",
                    "email_verified_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "users"
                    },
                    "customer_id": {
                        "nullable": "true",
                        "type": "integer",
                        "example": "null"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin Admins"
                    },
                    "email": {
                        "nullable": "true",
                        "type": "string",
                        "example": "admin@email.com"
                    },
                    "email_verified_at": {
                        "nullable": "true",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}